Move findDirWithFile to files.ts
